smart-agriculture-system/requirements.txt
flask
numpy
pandas
scikit-learn
matplotlib
seaborn
tensorflow
pillow
gunicorn
